🚚 refactor: Improve seeds folder structure
