+++ src/seeds/seeds.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { SeedsService } from './seeds.service';
-import { SeedsController } from './seeds.controller';
+import { SeedsService } from './services/seeds.service';
+import { SeedsController } from './controllers/seeds.controller';
 import { StatesModule } from '@States/states.module';
 
 @Module({
